feat: migrate appLoader.ts
